Move BitCommand to IBitCommand interface
